+++ src/script/token.rs
@@ -1,12 +1,12 @@
-use std::iter::Peekable;
+use std::{fmt::Display, iter::Peekable};
 
-use super::Error;
+use super::{parser::Parse, Error};
 
 #[cfg(test)]
 mod tests {
     use crate::{
         script::token::{
-            Eq, Ident, Let, NamedIdent, Number, PointCollection, Position, Semi, Span,
+            Dot, Eq, Ident, Let, NamedIdent, Number, PointCollection, Position, Semi, Span,
         },
         span,
     };
@@ -37,7 +37,8 @@ let ABC = Triangle;
                     span: span!(1, 9, 1, 10),
                     integral: 5,
                     decimal: 0,
-                    decimal_places: 0
+                    decimal_places: 0,
+                    dot: None
                 }),
                 Token::Semi(Semi {
                     span: span!(1, 10, 1, 11)
@@ -56,7 +57,10 @@ let ABC = Triangle;
                     span: span!(2, 9, 2, 12),
                     integral: 5,
                     decimal: 5,
-                    decimal_places: 2
+                    decimal_places: 2,
+                    dot: Some(Dot {
+                        span: span!(2, 10, 2, 11)
+                    })
                 }),
                 Token::Semi(Semi {
                     span: span!(2, 12, 2, 13)
@@ -99,13 +103,8 @@ impl PartialOrd for Position {
 impl Ord for Position {
     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
         match self.line.cmp(&other.line) {
-            std::cmp::Ordering::Less => std::cmp::Ordering::Less,
-            std::cmp::Ordering::Equal => match self.column.cmp(&other.column) {
-                std::cmp::Ordering::Less => std::cmp::Ordering::Less,
-                std::cmp::Ordering::Equal => std::cmp::Ordering::Equal,
-                std::cmp::Ordering::Greater => std::cmp::Ordering::Greater,
-            },
-            std::cmp::Ordering::Greater => std::cmp::Ordering::Greater,
+            std::cmp::Ordering::Equal => self.column.cmp(&other.column),
+            v => v,
         }
     }
 }
@@ -132,6 +131,32 @@ impl Span {
                 other.end
             },
         }
+    }
+
+    #[must_use]
+    pub fn overlaps(self, other: Span) -> bool {
+        (self.start <= other.start && self.end >= other.start)
+            || (other.start <= self.start && other.end >= self.start)
+    }
+
+    #[must_use]
+    pub fn is_singleline(&self) -> bool {
+        self.start.line == self.end.line
+    }
+}
+
+impl PartialOrd for Span {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        match self.start.cmp(&other.start) {
+            std::cmp::Ordering::Equal => Some(self.end.cmp(&other.end)),
+            v => Some(v),
+        }
+    }
+}
+
+impl Ord for Span {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.partial_cmp(other).unwrap()
     }
 }
 
@@ -235,6 +260,12 @@ pub struct Lteq {
     pub span: Span,
 }
 
+/// A '.' token.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Dot {
+    pub span: Span,
+}
+
 /// A '>=' token.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Gteq {
@@ -267,6 +298,77 @@ pub enum Token {
     Exclamation(Exclamation),
     Ident(Ident),
     Number(Number),
+}
+
+impl Display for Token {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Token::Semi(_) => write!(f, ";"),
+            Token::Eq(_) => write!(f, "="),
+            Token::Comma(_) => write!(f, ","),
+            Token::Let(_) => write!(f, "let"),
+            Token::Plus(_) => write!(f, "+"),
+            Token::Minus(_) => write!(f, "-"),
+            Token::Asterisk(_) => write!(f, "*"),
+            Token::Vertical(_) => write!(f, "|"),
+            Token::LParen(_) => write!(f, "("),
+            Token::RParen(_) => write!(f, ")"),
+            Token::Slash(_) => write!(f, "/"),
+            Token::Lt(_) => write!(f, "<"),
+            Token::Gt(_) => write!(f, ">"),
+            Token::Lteq(_) => write!(f, "<="),
+            Token::Gteq(_) => write!(f, ">="),
+            Token::Exclamation(_) => write!(f, "!"),
+            Token::Ident(ident) => write!(
+                f,
+                "{}",
+                match ident {
+                    Ident::Named(named) => named.ident.clone(),
+                    Ident::Collection(col) => col
+                        .collection
+                        .iter()
+                        .map(|(letter, primes)| format!("{letter}{}", "'".repeat(*primes as usize)))
+                        .collect::<String>(),
+                }
+            ),
+            Token::Number(num) => match num.dot {
+                Some(_) => write!(
+                    f,
+                    "{}.{:0>places$}",
+                    num.integral,
+                    num.decimal,
+                    places = num.decimal_places as usize
+                ),
+                None => write!(f, "{}", num.integral),
+            },
+        }
+    }
+}
+
+impl Token {
+    #[must_use]
+    pub fn get_span(&self) -> Span {
+        match self {
+            Token::Semi(v) => v.span,
+            Token::Eq(v) => v.span,
+            Token::Comma(v) => v.span,
+            Token::Let(v) => v.span,
+            Token::Plus(v) => v.span,
+            Token::Minus(v) => v.span,
+            Token::Asterisk(v) => v.span,
+            Token::Vertical(v) => v.span,
+            Token::LParen(v) => v.span,
+            Token::RParen(v) => v.span,
+            Token::Slash(v) => v.span,
+            Token::Lt(v) => v.span,
+            Token::Gt(v) => v.span,
+            Token::Lteq(v) => v.span,
+            Token::Gteq(v) => v.span,
+            Token::Exclamation(v) => v.span,
+            Token::Ident(v) => v.get_span(),
+            Token::Number(v) => v.span,
+        }
+    }
 }
 
 /// A name.
@@ -313,13 +415,14 @@ impl Ident {
     }
 }
 
-/// A 'ident' token.
+/// A number.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Number {
     pub span: Span,
     pub integral: u64,
     pub decimal: u64,
     pub decimal_places: u8,
+    pub dot: Option<Dot>,
 }
 
 fn is_identifier_character(c: char) -> bool {
@@ -358,6 +461,7 @@ fn read_number<I: Iterator<Item = char>>(it: &mut Peekable<I>, position: &mut Po
     let mut integral: u64 = 0;
     let mut decimal: u64 = 0;
     let begin_pos = *position;
+    let mut dot = None;
 
     while let Some(&c) = it.peek() {
         if c.is_ascii_digit() {
@@ -366,6 +470,14 @@ fn read_number<I: Iterator<Item = char>>(it: &mut Peekable<I>, position: &mut Po
             position.column += 1;
             it.next();
         } else if c == '.' {
+            dot = Some(Dot {
+                span: span!(
+                    position.line,
+                    position.column,
+                    position.line,
+                    position.column + 1
+                ),
+            });
             position.column += 1;
             it.next();
             break;
@@ -380,6 +492,7 @@ fn read_number<I: Iterator<Item = char>>(it: &mut Peekable<I>, position: &mut Po
                 integral,
                 decimal: 0,
                 decimal_places: 0,
+                dot,
             };
         }
     }
@@ -408,6 +521,7 @@ fn read_number<I: Iterator<Item = char>>(it: &mut Peekable<I>, position: &mut Po
         integral,
         decimal,
         decimal_places,
+        dot,
     }
 }
 
@@ -502,7 +616,7 @@ fn tokenize_special<I: Iterator<Item = char>>(
             '<' => Token::Lt(Lt { span: sp }),
             '>' => Token::Gt(Gt { span: sp }),
             '!' => Token::Exclamation(Exclamation { span: sp }),
-            _ => return Err(Error::invalid_character(c)),
+            _ => return Err(Error::invalid_character(c, sp)),
         });
     }