+++ changelog.md
@@ -7,5 +7,7 @@
 - Refactorized svg.rs to optimize the rendering
 - Created the latex drawer
 - Completed the documentation as of now.
+- Added a Command Line Interface
+- Pretty error printing
 - Added the .json drawer
 - Added the raw drawer